+++ CHANGELOG.rst
@@ -93,6 +93,20 @@ Release notes
   - Pull ScanCode.io Project data
   https://github.com/aboutcode-org/dejacode/issues/94
 
+- Add a new Vulnerabilities list available from the "Tools" menu when
+  ``enable_vulnerablecodedb_access`` is enabled on a Dataspace.
+  This implementation focuses on ranking/sorting: Vulnerabilities can be sorted and
+  filtered by severity score.
+  It's also possible to sort by the count of affected packages to help prioritize.
+  https://github.com/aboutcode-org/dejacode/issues/94
+
+- Display warning when a "download_url" could not be determined from a PURL in
+  "Add Package".
+  https://github.com/aboutcode-org/dejacode/issues/163
+
+- Add a Vulnerabilities tab in the Product details view.
+  https://github.com/aboutcode-org/dejacode/issues/95
+
 ### Version 5.1.0
 
 - Upgrade Python version to 3.12 and Django to 5.0.x